refactor: benchmark runner
